+++ changelog.md
@@ -7,7 +7,7 @@ This version is not released yet and is under active development.
 ```
 
 - Expose location and content of user's configuration file in the Context's `meta` property. Closes {issue}`673`.
-- Fix display of hidden parameters in `--show-params` output. Fixes {issue}`689`.
+- Render specs of hidden parameters in `--show-params` output. Fixes {issue}`689`.
 - Swap `Exposed` and `Allowed in conf?` columns in `--show-params` output.
 - Add a `hidden` column to `--show-params` output. Refs {issue}`689`.